futureagi/pyproject.toml
[project]
name = "core-backend"
version = "0.1.0"
description = "Core backend for FutureAGI platform"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django==5.1.8",
    "djangorestframework",
    "djangorestframework-simplejwt",
    "granian[uvloop,reload]",  # Rust ASGI server
    "psycopg[binary,pool]",
    "django-anymail",
    "django-cors-headers",

    "clickhouse-driver",
    "openai>=2.15.0",  # Latest with pydantic v2.11 compatibility
    "multidict>=6.7.0",  # Fix memory corruption in 6.6.0 (yanked)
    "watchdog",
    "watchfiles",  # Rust-based file watcher (faster, works in Docker)
    "tenacity",
    "boto3",
    "celery",
    "django-celery-results",
    "drf-yasg",
    "django-debug-toolbar",
    "slack-sdk",
    "chromadb",
    "flower",
    "pydantic>=2.11.10",  # Use latest with thread-safety fixes
    "jsonpath-ng",
    "nltk",
    "retrying",
    "litellm==1.81.11",  # Latest with pydantic compatibility fixes
    "django-filter",
    "pyjwt",
    "qdrant-client",
    "weaviate-client",
    "stripe",
    # OpenTelemetry core
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-proto",
    "opentelemetry-semantic-conventions",
    # OpenTelemetry propagators (for context propagation between spans/services)
    "opentelemetry-propagator-b3>=1.20.0",
    # OpenTelemetry exporters
    "opentelemetry-exporter-otlp>=1.20.0",
    "opentelemetry-exporter-otlp-proto-common",
    "opentelemetry-exporter-otlp-proto-grpc",
    "opentelemetry-exporter-otlp-proto-http",
    # OpenTelemetry instrumentation - core
    "opentelemetry-instrumentation>=0.41b0",
    "opentelemetry-util-http",
    # OpenTelemetry instrumentation - Django
    "opentelemetry-instrumentation-django>=0.41b0",
    "opentelemetry-instrumentation-asgi",
    # OpenTelemetry instrumentation - Celery
    "opentelemetry-instrumentation-celery>=0.41b0",
    # OpenTelemetry instrumentation - HTTP clients
    "opentelemetry-instrumentation-requests>=0.41b0",
    "opentelemetry-instrumentation-httpx>=0.41b0",
    # OpenTelemetry instrumentation - gRPC
    "opentelemetry-instrumentation-grpc>=0.41b0",
    # OpenTelemetry instrumentation - Database
    "opentelemetry-instrumentation-psycopg>=0.41b0",
    "opentelemetry-instrumentation-redis>=0.41b0",
    # OpenTelemetry instrumentation - HTTP (additional)
    "opentelemetry-instrumentation-urllib3>=0.41b0",
    # OpenTelemetry instrumentation - Logging (for structlog correlation)
    "opentelemetry-instrumentation-logging>=0.41b0",
    # OpenTelemetry instrumentation - Templates
    "opentelemetry-instrumentation-jinja2>=0.41b0",
    # OpenTelemetry instrumentation - aiohttp (async HTTP client)
    "opentelemetry-instrumentation-aiohttp-client>=0.41b0",
    # OpenTelemetry instrumentation - AWS (botocore/boto3)
    "opentelemetry-instrumentation-botocore>=0.41b0",
    # traceAI - OpenTelemetry instrumentation for LLMs (https://github.com/future-agi/traceAI)
    "traceai-openai",
    "traceai-anthropic",
    "traceai-bedrock",
    "traceai-google-genai",
    "traceai-vertexai",
    "traceai-litellm",
    # OpenTelemetry instrumentation - messaging
    "opentelemetry-instrumentation-pika>=0.41b0",
    # OpenTelemetry instrumentation - asyncio
    "opentelemetry-instrumentation-asyncio>=0.41b0",
    "protobuf",
    "xmlsec",
    "pysaml2",
    "python-jose",
    "pinecone==5.4.2",
    "mixpanel",
    "posthog>=7.9.0",
    "minio",
    "django-celery-beat",
    "channels",
    "channels-rabbitmq",
    "pydub",
    "filetype",
    "google-genai",
    "ffmpeg-python",
    "numba>=0.60.0",
    "llvmlite>=0.43.0",
    "librosa>=0.10.1",
    "soundfile",
    "json-repair",
    "pdfminer.six",
    "python-docx",
    "striprtf",
    "pdfminer",
    "openpyxl>=3.1.0",
    "pdfplumber>=0.10.0",
    "beautifulsoup4>=4.12.0",
    "partial-json-parser",
    "chevron",
    "jinja2",
    "pypdf",
    "google-cloud-aiplatform",
    "anthropic[vertex]",
    "rouge-score",
    "python-levenshtein",
    "scipy",
    "numpy",
    "torch",
    "torchvision",
    "torchmetrics",
    "torch-fidelity",
    "sentry-sdk",
    "structlog>=24.1.0",
    "django-structlog>=8.1.0",
    "django-socio-grpc",
    "av>=14.0.0,!=14.4.0,<15.0.0",  # 14.4.0 has no wheels; use 14.3.0 or similar
    "datasets",
    "pillow",
    "pandas",
    "langchain",
    "transformers",
    "langchain_text_splitters",
    "langchain_community",
    "langfuse==3.1.3",
    "whitenoise",
    "hdbscan",
    "elevenlabs",
    "retell-sdk",
    "temporalio[opentelemetry]",  # opentelemetry extra enables TracingInterceptor
    "gepa==0.1.1",
    "optuna>=4.6.0",
    "redis>=5.0.0",  # For distributed locking and state management
    "django-redis>=5.4.0",  # Redis cache backend
    "livekit-api>=1.0.0",
    "livekit>=1.0.0",
    # 2FA & Passkeys
    "pyotp>=2.9.0",  # TOTP generation and verification
    "qrcode[pil]>=7.4",  # QR code generation for TOTP provisioning URIs
    "webauthn>=2.2.0",  # WebAuthn server-side (py_webauthn by Duo Labs)
    "django-stubs>=5.2.8",
    "djangorestframework-stubs>=3.16.7",
    "presidio-analyzer>=2.2.0",  # PII Redaction
    "presidio-anonymizer>=2.2.0",
    "spacy>=3.7.0",
    "mcp>=1.26.0",  # Model Context Protocol SDK (Anthropic)
]

[project.optional-dependencies]
dev = [
    "grpcio",
    "grpcio-tools",
    "pre-commit",
    "mypy>=1.8.0",
    "django-stubs>=4.2.7",
    "djangorestframework-stubs>=3.14.5",
    "types-requests",
    "types-redis",
    "types-pyyaml",
]

[tool.black]
line-length = 88
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 88
known_third_party = ["your_third_party_packages"]
known_first_party = ["your_project_name"]
multi_line_output = 3
include_trailing_comma = true
use_parentheses = true

[tool.bandit]
exclude_dirs = ["tests", "*/tests/*", "agentic_eval/tests"]
skips = ["B101", "B601"]  # B101: assert_used, B601: paramiko_calls
# Uncomment to skip specific tests:
# tests = []

[tool.ruff]
line-length = 88
target-version = "py311"
exclude = [
    ".git",
    "__pycache__",
    "migrations",
    "*.pyc",
    ".pytest_cache",
    "htmlcov",
    "static",
    "media",
]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports

[tool.mypy]
python_version = "3.11"
plugins = [
    "mypy_django_plugin.main",
    "mypy_drf_plugin.main",
]

# Strict settings
strict_optional = true
no_implicit_optional = true
warn_unused_ignores = true
check_untyped_defs = true
warn_unreachable = true
strict_equality = true
warn_return_any = true
warn_unused_configs = true

# Still ignore missing imports for now (can be tightened later)
ignore_missing_imports = true

# Exclude patterns
exclude = [
    "migrations/",
    "tests/",
    "agentic_eval/",
    "__pycache__/",
]

[tool.django-stubs]
django_settings_module = "tfc.settings"

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=6.2.1",
    "pytest-django>=4.11.1",
    "pytest-mock>=3.14.1",
    "pytest-xdist>=3.8.0",
]
